requirements.txt
requests
tornado
pyyaml